requirements.txt
aiogram
aiohttp
python-dotenv
openai
fastapi
uvicorn
pydantic
RestrictedPython